Level 2 adjustments & git config
Spun level 2 around the proper way, moved it around a bit with respect to the terrain.   Added a .gitattributes file that will override client settings and ensure that .pbt files can use LF instead of CRLF for line endings:  This appears to be the way that the Core engine likes things.