Golden path work
I lowered the town to the ground-ish.  The effect is very nice inside the tavern, IMHO, if a bit crowded by trees.  Oh, and there's only one way in/out, it seems :P  No bridge over the fire:  Did not need one ;)  Got to the game board and died right quick.  We need a weapons and potions store right quick.  Where is the boss going to live?