Level 2 and game changes
Placed gate key and loot for level 2.  Scaled the exit gate down a bit, to fit the wall height.  Built new wall around level to force players to go through the board entrance.  Changed chat mode to "all" for co-op play.  Disabled mounts, since they can be used to cheat the overall difficulty of the game, and in multiple levels.